build: modify deployment-rest.yaml and directory name  (#2)
