--- src/cloudforet/console_api/__init__.py
@@ -1,2 +1,0 @@
-name = 'console_api'
-
