--- src/cloudforet/console_api/error/__init__.py
@@ -1,1 +1,0 @@
-from cloudforet.console_api.error.cloudforet import *
