+++ src/cloudforet/console_api_v2/__init__.py
@@ -1,0 +1,2 @@
+name = 'console_api_v2'
+
