+++ src/cloudforet/console_api_v2/conf/router_conf.py
@@ -1,16 +1,16 @@
 ROUTER = [
     {
-        'path': 'cloudforet.console_api.interface.rest.dashboard.domain_dashboard:router',
+        'path': 'cloudforet.console_api_v2.interface.rest.dashboard.domain_dashboard:router',
         'prefix': '/dashboard/domain-dashboard',
         'tags': ['dashboard > domain-dashboard'],
     },
     {
-        'path': 'cloudforet.console_api.interface.rest.dashboard.project_dashboard:router',
+        'path': 'cloudforet.console_api_v2.interface.rest.dashboard.project_dashboard:router',
         'prefix': '/dashboard/project-dashboard',
         'tags': ['dashboard > project-dashboard'],
     },
     {
-        'path': 'cloudforet.console_api.interface.rest.dashboard.custom_widget:router',
+        'path': 'cloudforet.console_api_v2.interface.rest.dashboard.custom_widget:router',
         'prefix': '/dashboard/custom-widget',
         'tags': ['dashboard > custom-widget'],
     }
