+++ src/cloudforet/console_api_v2/error/__init__.py
@@ -1,0 +1,1 @@
+from cloudforet.console_api_v2.error.cloudforet import *
