+++ src/cloudforet/console_api_v2/interface/rest/dashboard/custom_widget.py
@@ -4,8 +4,8 @@ from fastapi_utils.inferring_router import InferringRouter
 from fastapi_utils.cbv import cbv
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from spaceone.core.fastapi.api import BaseAPI, exception_handler
-from cloudforet.console_api.model.dashboard.custom_widget import *
-from cloudforet.console_api.service.common.proxy_service import ProxyService
+from cloudforet.console_api_v2.model.dashboard.custom_widget import *
+from cloudforet.console_api_v2.service.common.proxy_service import ProxyService
 
 _LOGGER = logging.getLogger(__name__)
 _AUTH_SCHEME = HTTPBearer()
