+++ src/cloudforet/console_api_v2/interface/rest/dashboard/domain_dashboard.py
@@ -3,10 +3,9 @@ from fastapi import Request, Depends
 from fastapi_utils.inferring_router import InferringRouter
 from fastapi_utils.cbv import cbv
 from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
-
 from spaceone.core.fastapi.api import BaseAPI, exception_handler
-from cloudforet.console_api.model.dashboard.project_dashboard import *
-from cloudforet.console_api.service.common.proxy_service import ProxyService
+from cloudforet.console_api_v2.model.dashboard.domain_dashboard import *
+from cloudforet.console_api_v2.service.common.proxy_service import ProxyService
 
 _LOGGER = logging.getLogger(__name__)
 _AUTH_SCHEME = HTTPBearer()
@@ -15,117 +14,126 @@ router = InferringRouter()
 
 
 @cbv(router)
-class ProjectDashboard(BaseAPI):
+class DomainDashboard(BaseAPI):
+
     token: HTTPAuthorizationCredentials = Depends(_AUTH_SCHEME)
     service = 'console-api'
 
     @router.post(
         '/create',
+        description=DomainDashboardInfo.Create.description(),
+        responses=DomainDashboardInfo.Create.response()
     )
     @exception_handler
-    async def create(self, request: Request, body: ProjectDashboardInfo.Create):
+    async def create(self, request: Request, body: DomainDashboardInfo.Create):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.create'
+            params['grpc_method'] = 'dashboard.DomainDashboard.create'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/update'
+        '/update',
+        description=DomainDashboardInfo.Update.description(),
+        responses=DomainDashboardInfo.Update.response(),
     )
     @exception_handler
-    async def update(self, request: Request, body: ProjectDashboardInfo.Update):
+    async def update(self, request: Request, body: DomainDashboardInfo.Update):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.update'
+            params['grpc_method'] = 'dashboard.DomainDashboard.update'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/delete'
+        '/delete',
+        description=DomainDashboardInfo.Delete.description()
     )
-    @exception_handler
-    async def delete(self, request: Request, body: ProjectDashboardInfo.Delete):
+    async def delete(self, request: Request, body: DomainDashboardInfo.Delete):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.delete'
+            params['grpc_method'] = 'dashboard.DomainDashboard.delete'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/get'
+        '/get',
+        description=DomainDashboardInfo.Get.description(),
+        responses=DomainDashboardInfo.Get.response()
     )
-    @exception_handler
-    async def get(self, request: Request, body: ProjectDashboardInfo.Get):
+    async def get(self, request: Request, body: DomainDashboardInfo.Get):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.get'
+            params['grpc_method'] = 'dashboard.DomainDashboard.get'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/delete_version'
+        '/delete_version',
     )
-    @exception_handler
-    async def delete_version(self, request: Request, body: ProjectDashboardVersionInfo.DeleteVersion):
+    async def delete_version(self, request: Request, body: DomainDashboardVersionInfo.DeleteVersion):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.delete_version'
+            params['grpc_method'] = 'dashboard.DomainDashboard.delete_version'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/revert_version'
+        '/revert_version',
+        description=DomainDashboardVersionInfo.RevertVersion.description(),
+        responses=DomainDashboardVersionInfo.RevertVersion.response()
     )
-    @exception_handler
-    async def revert_version(self, request: Request, body: ProjectDashboardVersionInfo.RevertVersion):
+    async def revert_version(self, request: Request, body: DomainDashboardVersionInfo.RevertVersion):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.revert_version'
+            params['grpc_method'] = 'dashboard.DomainDashboard.revert_version'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/get_version'
+        '/get_version',
+        description=DomainDashboardVersionInfo.GetVersion.description(),
+        responses=DomainDashboardVersionInfo.GetVersion.response()
     )
-    @exception_handler
-    async def get_version(self, request: Request, body: ProjectDashboardVersionInfo.RevertVersion):
+    async def get_version(self, request: Request, body: DomainDashboardVersionInfo.GetVersion):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.get_version'
+            params['grpc_method'] = 'dashboard.DomainDashboard.get_version'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/list_versions'
+        '/list_versions',
+        description=DomainDashboardVersionInfo.ListVersions.description(),
+        responses=DomainDashboardVersionInfo.ListVersions.response()
     )
-    @exception_handler
-    async def list_versions(self, request: Request, body: ProjectDashboardVersionInfo.ListVersions):
+    async def list_versions(self, request: Request, body: DomainDashboardVersionInfo.ListVersions):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.list_versions'
+            params['grpc_method'] = 'dashboard.DomainDashboard.list_versions'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/list'
+        '/list',
+        description=DomainDashboardInfo.List.description(),
+        responses=DomainDashboardInfo.List.response()
     )
-    @exception_handler
-    async def list(self, request: Request, body: ProjectDashboardInfo.List):
+    async def list(self, request: Request, body: DomainDashboardInfo.List):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.list'
+            params['grpc_method'] = 'dashboard.DomainDashboard.list'
             return proxy_service.dispatch_api(params)
 
     @router.post(
-        '/stat'
+        '/stat',
+        description=Stat.description(),
+        responses=Stat.response()
     )
-    @exception_handler
     async def stat(self, request: Request, body: Stat):
         params, metadata = await self.parse_request(request, self.token.credentials)
 
         with self.locator.get_service(ProxyService, metadata) as proxy_service:
-            params['grpc_method'] = 'dashboard.ProjectDashboard.stat'
+            params['grpc_method'] = 'dashboard.DomainDashboard.stat'
             return proxy_service.dispatch_api(params)
-
