+++ src/cloudforet/console_api_v2/model/dashboard/custom_widget.py
